third_party/blink/web_tests/wpt_internal/css/css-overflow/scrollbar-gutter-004.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Overflow: test scrollbar-gutter: auto (classic scrollbars, vertical text)</title>
<link rel="author" title="Felipe Erias Morandeira" href="mailto:felipeerias@gmail.com"/>
<link rel="help" href="https://www.w3.org/TR/css-overflow-4/#scollbar-gutter-property"/>
<meta name="assert" content="Test scrollbar-gutter: auto with custom classic scrollbars">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<style>

    .container {
        scrollbar-gutter: auto;
        height: 200px;
        width: 200px;
        margin: 1px;
        padding: 0px;
        border: none;
        background: deepskyblue;
    }

    .content {
        height: 100%;
        width: 100%;
        background: lightsalmon;
    }

    /* ensure that we are using classic scrollbars with a known thickness */
    .classic::-webkit-scrollbar {
        width: 20px;
        height: 20px;
    }
    .classic::-webkit-scrollbar-track {
        background-color: hsla(0, 0%, 40%, 0.4);
    }
    .classic::-webkit-scrollbar-thumb {
        background-color: hsla(0, 0%, 20%, 0.8);
        border-radius: 10px;
    }

    /* writing mode */
    .vertical {
        writing-mode: vertical-rl;
        direction: ltr;
        overflow-y: auto;
    }

    /* overflow on the block direction */
    .vertical.auto      { overflow-x: auto;    }
    .vertical.scroll    { overflow-x: scroll;  }
    .vertical.visible   { overflow:   visible; }
    .vertical.hidden    { overflow-x: hidden;  }

    /* Note: not testing with overflow: clip; */

</style>
<script type="text/javascript">

    function performTest() {
        setup({explicit_done: true});

        test(function() {
            let container = document.getElementById('container_auto_v');
            let content = document.getElementById('content_auto_v');
            assert_equals(container.scrollHeight, 200, "auto/auto scrollHeight");
            assert_equals(container.clientHeight, 200, "auto/auto clientHeight");
            assert_equals(container.offsetTop, content.offsetTop, "auto/auto offsetTop");
            assert_equals(container.clientHeight, content.clientHeight, "auto/auto clientHeight");
            assert_equals(container.offsetHeight, content.offsetHeight, "auto/auto offsetHeight");
        }, "overflow auto, scrollbar-gutter auto, vertical-rl");

        test(function() {
            let container = document.getElementById('container_scroll_v');
            let content = document.getElementById('content_scroll_v');
            assert_equals(container.scrollHeight, 180, "scroll/auto scrollHeight");
            assert_equals(container.clientHeight, 180, "scroll/auto clientHeight");
            assert_equals(container.offsetTop, content.offsetTop, "scroll/auto offsetTop");
            assert_equals(container.clientHeight, content.clientHeight, "scroll/auto clientHeight");
            assert_not_equals(container.offsetHeight, content.offsetHeight, "scroll/auto offsetHeight");
        }, "overflow scroll, scrollbar-gutter auto, vertical-rl");

        test(function() {
            let container = document.getElementById('container_visible_v');
            let content = document.getElementById('content_visible_v');
            assert_equals(container.scrollHeight, 200, "visible/auto scrollHeight");
            assert_equals(container.clientHeight, 200, "visible/auto clientHeight");
            assert_equals(container.offsetTop, content.offsetTop, "visible/auto offsetTop");
            assert_equals(container.clientHeight, content.clientHeight, "visible/auto clientHeight");
            assert_equals(container.offsetHeight, content.offsetHeight, "visible/auto offsetHeight");
        }, "overflow visible, scrollbar-gutter auto, vertical-rl");

        test(function() {
            let container = document.getElementById('container_hidden_v');
            let content = document.getElementById('content_hidden_v');
            assert_equals(container.scrollHeight, 200, "hidden/auto scrollHeight");
            assert_equals(container.clientHeight, 200, "hidden/auto clientHeight");
            assert_equals(container.offsetTop, content.offsetTop, "hidden/auto offsetTop");
            assert_equals(container.clientHeight, content.clientHeight, "hidden/auto clientHeight");
            assert_equals(container.offsetHeight, content.offsetHeight, "hidden/auto offsetHeight");
        }, "overflow hidden, scrollbar-gutter auto, vertical-rl");

        done();
    }

</script>
<body onload="performTest()">

    Test scrollbar-gutter: auto, writing-mode: vertical-rl

    <div class="container classic vertical auto" id="container_auto_v">
        <div class="content" id="content_auto_v">ｏｖｅｒｆｌｏｗ：ａｕｔｏ</div>
    </div>

    <div class="container classic vertical scroll" id="container_scroll_v">
        <div class="content" id="content_scroll_v">ｏｖｅｒｆｌｏｗ－ｘ：ｓｃｒｏｌｌ</div>
    </div>

    <div class="container classic vertical visible" id="container_visible_v">
        <div class="content" id="content_visible_v">ｏｖｅｒｆｌｏｗ：ｖｉｓｉｂｌｅ</div>
    </div>

    <div class="container classic vertical hidden" id="container_hidden_v">
        <div class="content" id="content_hidden_v">ｏｖｅｒｆｌｏｗ－ｘ：ｈｉｄｄｅｎ</div>
    </div>

</body>
